chore: move MeshInfo2Holder to another file
